refactor: 사물함 페이지를 widgets와 features 중심으로 재구성 #53
