--- apps/web/src/_pages/locker/model/index.ts
@@ -1,4 +1,0 @@
-export * from './lib';
-export * from './types';
-export * from './useLockerListPage';
-export * from './useLockerToast';
