+++ apps/web/src/widgets/locker/index.ts
@@ -1,0 +1,1 @@
+export { LockerPageWidget } from './ui';
