+++ apps/web/src/widgets/locker/model/index.ts
@@ -1,0 +1,3 @@
+export * from './lib';
+export * from './types';
+export * from './useLockerPage';
