+++ apps/web/src/widgets/locker/ui/index.ts
@@ -1,4 +1,5 @@
 export * from './FloorSummaryCard';
 export * from './LockerInfoCard';
 export * from './LockerNoticeCard';
+export * from './LockerPageWidget';
 export * from './LockerSelectionOverlay';
